refactor(codex): isolate JSONL line scanning
